requirements.txt
bump2version==1.0.1
coverage==6.5.0
Django==4.1.3
django-environ==0.9.0
django-redis==5.2.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.2.1
drf-spectacular==0.24.2
fakeredis==2.2.0
flake8==6.0.0
flake8-isort==5.0.3
isort==5.10.1
mongoengine==0.24.2
mongomock==4.1.2
pymongo==4.3.3
pytest==7.2.0
pytest-django==4.5.2
pytest-cov==4.0.0
safety==2.3.3